refactor: gnucash-schema Row types
